Refactor social feed into smaller components
